New "repeat" module (#23)
* Reimplement and rename get_annotation function

* loci_need_split_anno --> collapse_repeats_by_length

* Fix style violations

* split_string --> sequence_to_bracketed_form

* New repeat module
