+++ lusSTR/__init__.py
@@ -8,6 +8,7 @@
 # -----------------------------------------------------------------------------
 
 from lusSTR import annot
+from lusSTR import repeat
 from lusSTR.annot import str_dict
 from lusSTR import format
 from lusSTR import cli
