+++ lusSTR/annot.py
@@ -14,7 +14,10 @@ import pandas as pd
 from pkg_resources import resource_filename
 import re
 import sys
+
 import lusSTR
+from lusSTR.repeat import collapse_all_repeats, collapse_repeats_by_length
+from lusSTR.repeat import sequence_to_bracketed_form, split_by_n
 
 
 def get_str_metadata_file():
@@ -23,73 +26,6 @@ def get_str_metadata_file():
 
 with open(get_str_metadata_file(), 'r') as fh:
     str_dict = json.load(fh)
-
-
-def collapse_tandem_repeat(fullseq, repeat):
-    '''Collapse tandem stretches of the specified repeat sequence in a larger sequence.
-
-    >>> collapse_tandem_repeat('TAGATTATTATTTAGTAGATTTAGTAG', 'ATT')
-    'TAG [ATT]3 TAGTAG ATT TAGTAG'
-    >>> collapse_tandem_repeat('TAGATTATTATTTAGTAGATTTAGTAG', 'TAG')
-    'TAG ATTATTATT [TAG]2 ATT [TAG]2'
-    '''
-    if repeat not in fullseq:
-        return fullseq
-    i = fullseq.find(repeat)
-    prefix = fullseq[:i]
-    suffix = fullseq[i:]
-    count = 0
-    while suffix.startswith(repeat):
-        count += 1
-        suffix = suffix[len(repeat):]
-    if count == 1:
-        formatted = f' {repeat} '
-    else:
-        formatted = f' [{repeat}]{count} '
-    final = prefix + formatted + collapse_tandem_repeat(suffix, repeat)
-    final = final.strip()
-    final = re.sub(r' +', ' ', final)
-    return final
-
-
-def collapse_all_repeats(sequence, repeats):
-    '''Convert a sequence to bracketed form by collapsing stretches of tandem repeats.
-
-    >>> collapse_all_repeats('TAGATTATTATTTAGTAGATTTAGTAG', ['ATT', 'TAG'])
-    'TAG [ATT]3 [TAG]2 ATT [TAG]2'
-    '''
-    collapsed_seq = sequence
-    for repeat in repeats:
-        collapsed_seq = collapse_tandem_repeat(collapsed_seq, repeat)
-    return collapsed_seq
-
-
-def split_by_n(sequence, n):
-    '''
-    Function to divide sequence into chunks of n
-    '''
-    while sequence:
-        yield sequence[:n]
-        sequence = sequence[n:]
-
-
-def sequence_to_bracketed_form(sequence, n, repeats):
-    '''Convert sequence to bracketed annotation.
-
-    Uses a combination of repeat-based and length-based methods to convert a sequence containing
-    tandem repeats into a concise bracketed representation.
-    '''
-    collapsed = collapse_all_repeats(sequence, repeats)
-    blocks = list()
-    for unit in collapsed.split(' '):
-        if len(unit) > n and '[' not in unit:
-            for x in split_by_n(unit, n):
-                blocks.append(x)
-        else:
-            blocks.append(unit)
-    result = ' '.join(blocks)
-    result = re.sub(r' +', ' ', result)
-    return result
 
 
 def rev_complement_anno(sequence):
@@ -171,38 +107,6 @@ def rev_comp_uas_output_bracket(forward_bracket, n):
             reverse_form_anno_final.append(unit)
     reverse_strand_bracketed_form = ' '.join(reverse_form_anno_final)
     return re.sub('  ', ' ', reverse_strand_bracketed_form)
-
-
-def get_blocks(sequence, n):
-    '''
-    Function to split a sequence into blocks of size n
-
-    This function is used as a part of the collapse_repeats_by_length() function. It splits the
-    sequence into blocks of size n bases (as specified in the str_markers.json file).
-    '''
-    count = 0
-    prev = None
-    for unit in split_by_n(sequence, n):
-        if unit != prev:
-            if prev is not None:
-                yield prev, count
-            prev = unit
-            count = 0
-        count += 1
-    yield prev, count
-
-
-def collapse_repeats_by_length(sequence, n):
-    '''Convert to bracketed annotation form by splitting the sequence into blocks of size n.'''
-    units = list()
-    for unit, count in get_blocks(sequence, n):
-        if count == 1:
-            units.append(unit)
-        else:
-            units.append(f'[{unit}]{count}')
-    result = '  '.join(units)
-    result = re.sub(r' +', ' ', result)
-    return result
 
 
 def traditional_str_allele(sequence, n, n_sub_out):
