+++ lusSTR/tests/test_suite.py
@@ -27,40 +27,6 @@ def test_format():
         assert filecmp.cmp(formatoutput, outfile.name) is True
 
 
-@pytest.mark.parametrize('sequence, repeat_list, output', [
-    (
-        'AGACAGATAGATAGATAGATAGATAGATAGATAGATAGATAGATAGATAGACAGACAGACAGACAGACAGACAGAT',
-        ['AGAT', 'AGAC'], 'AGAC [AGAT]11 [AGAC]6 AGAT'
-    ),
-    (
-        'TAGATAGATAGATAGATGATAGATAGATAGATAGATAGATAGATAGATAGATAGATAGATAGATAGATAGATAGATAGATAGG',
-        ['TCTA', 'CATA', 'TCTG', 'CACA', 'CCTA'],
-        'TAGATAGATAGATAGATGATAGATAGATAGATAGATAGATAGATAGATAGATAGATAGATAGATAGATAGATAGATAGATAGG'
-    )
-])
-def test_collapse_all_repeats(sequence, repeat_list, output):
-    final_output = lusSTR.annot.collapse_all_repeats(sequence, repeat_list)
-    assert final_output == output
-
-
-def test_split_by_n():
-    sequence = 'AGGTAGGTAGGTCGAACGAATTGG'
-    blocks = list(lusSTR.annot.split_by_n(sequence, n=4))
-    assert blocks == [
-        'AGGT', 'AGGT', 'AGGT', 'CGAA', 'CGAA', 'TTGG'
-    ]
-
-
-def test_sequence_to_bracketed_form():
-    sequence = (
-        'TCTATCTATCTATCTGTCTGTCTGTCTGTCTGTCTGTCTATCTATCTATATCTATCTATCTATCATCTATCTATCCATATCTATCTATC'
-        'TATCTATCTATCTATCTATCTATCTATCTATCTA'
-    )
-    repeats = ['TCTA', 'TCTG']
-    final_output = lusSTR.annot.sequence_to_bracketed_form(sequence, 6, repeats)
-    assert final_output == '[TCTA]3 [TCTG]6 [TCTA]3 TA [TCTA]3 TCA [TCTA]2 TCCATA [TCTA]11'
-
-
 def test_extract():
     s = '[ATCT]3 ATGT [ATCT]12'
     repeat = 'ATCT'
@@ -90,11 +56,6 @@ def test_rev_comp_uas_output_bracket():
     foward_strand = '[AGGT]3 [CGAA]2 TTGG'
     rev_comp_bracket = lusSTR.annot.rev_comp_uas_output_bracket(foward_strand, 4)
     assert rev_comp_bracket == 'CCAA [TTCG]2 [ACCT]3'
-
-
-def test_collapse_repeats_by_length():
-    sequence = 'TCTATCTATCTATCTATCTATCTATCTATATATCTATCTATCTATCTA'
-    assert lusSTR.annot.collapse_repeats_by_length(sequence, 4) == '[TCTA]7 TATA [TCTA]4'
 
 
 @pytest.mark.parametrize('sequence, bracket_form', [
